remove unused tiled websoecket
